Move beam_generic code into beam.cpp
Cleanup beam.cpp and delete beam_generic.
Also, fix bug #560.
